changed to correct version, no issues now
