+++ DataBaseAssigmentConsoleApp/DataBaseAssigmentConsoleApp.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33213.308
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataBaseAssigmentConsoleApp", "DataBaseAssigmentConsoleApp\DataBaseAssigmentConsoleApp.csproj", "{1D9CC32D-E82C-4417-BE06-F3626A23B706}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1D9CC32D-E82C-4417-BE06-F3626A23B706}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D9CC32D-E82C-4417-BE06-F3626A23B706}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D9CC32D-E82C-4417-BE06-F3626A23B706}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D9CC32D-E82C-4417-BE06-F3626A23B706}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E949F3C3-EC2A-468F-822D-ADC4C80D71FB}
+	EndGlobalSection
+EndGlobal
