+++ DataBaseAssigmentConsoleApp/DataBaseAssigmentConsoleApp/Models/Customer.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DataBaseAssigmentConsoleApp.Models
+{
+        public class Customer
+        {
+            public int CustomerId { get; set; }
+            public string CompanyName { get; set; }
+            public string ContactName { get; set; }
+            public string City { get; set; }
+
+        }
+}
