+++ DataBaseAssigmentConsoleApp/DataBaseAssigmentConsoleApp/Program.cs
@@ -1,0 +1,10 @@
+﻿namespace DataBaseAssigmentConsoleApp
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello, World!");
+        }
+    }
+}
