+++ DataBaseAssigmentConsoleApp/DataBaseAssigmentConsoleApp/Repositories/ConnectionStringHelper.cs
@@ -1,11 +1,11 @@
-﻿using System;
+﻿using Microsoft.Data.SqlClient;
+using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabasesAssignmentTwoConsoleApp.Repositories
+namespace DataBaseAssigmentConsoleApp.Repositories
 {
     public class ConnectionStringHelper
     {
@@ -14,9 +14,8 @@ namespace DatabasesAssignmentTwoConsoleApp.Repositories
             SqlConnectionStringBuilder connectionStringBuilder = new SqlConnectionStringBuilder();
             connectionStringBuilder.DataSource = "N-NO-01-01-4564\\SQLEXPRESS";
             connectionStringBuilder.InitialCatalog = "Chinook";
-            connectionStringBuilder.IntegratedSecurity= true;
+            connectionStringBuilder.IntegratedSecurity = true;
             return connectionStringBuilder.ConnectionString;
-
         }
     }
 }
