--- DatabasesAssignmentTwoConsoleApp/DatabasesAssignmentTwoConsoleApp.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33213.308
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DatabasesAssignmentTwoConsoleApp", "DatabasesAssignmentTwoConsoleApp\DatabasesAssignmentTwoConsoleApp.csproj", "{DA3065F1-865D-43CA-B113-D8CED0E85BD5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DA3065F1-865D-43CA-B113-D8CED0E85BD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DA3065F1-865D-43CA-B113-D8CED0E85BD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DA3065F1-865D-43CA-B113-D8CED0E85BD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DA3065F1-865D-43CA-B113-D8CED0E85BD5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E059D304-99C7-462A-BB9C-5C7EC31720F6}
-	EndGlobalSection
-EndGlobal
