--- DatabasesAssignmentTwoConsoleApp/DatabasesAssignmentTwoConsoleApp/Models/Customer.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DatabasesAssignmentTwoConsoleApp.Models
-{
-    internal class Customer
-    {
-        public int CustomerId { get; set; }
-        public string CompanyName { get; set; }
-        public string ContactName { get; set; }
-        public string City { get; set; }
-
-    }
-}
