--- DatabasesAssignmentTwoConsoleApp/DatabasesAssignmentTwoConsoleApp/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DatabasesAssignmentTwoConsoleApp
-{
-    internal class Program
-    {
-        //Get all customers
-        //get 1 customer
-        //CRUD 
-    }
-}
