--- DatabasesAssignmentTwoConsoleApp/DatabasesAssignmentTwoConsoleApp/packages.config
@@ -1,41 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Azure.Core" version="1.25.0" targetFramework="net472" />
-  <package id="Azure.Identity" version="1.7.0" targetFramework="net472" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.1" targetFramework="net472" />
-  <package id="Microsoft.Data.SqlClient" version="5.1.0" targetFramework="net472" />
-  <package id="Microsoft.Data.SqlClient.SNI" version="5.1.0" targetFramework="net472" />
-  <package id="Microsoft.Identity.Client" version="4.47.2" targetFramework="net472" />
-  <package id="Microsoft.Identity.Client.Extensions.Msal" version="2.19.3" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Abstractions" version="6.24.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.JsonWebTokens" version="6.24.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Logging" version="6.24.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Protocols" version="6.24.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Protocols.OpenIdConnect" version="6.24.0" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Tokens" version="6.24.0" targetFramework="net472" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
-  <package id="System.Configuration.ConfigurationManager" version="6.0.1" targetFramework="net472" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.6.0" targetFramework="net472" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="6.24.0" targetFramework="net472" />
-  <package id="System.IO" version="4.3.0" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net472" />
-  <package id="System.Memory.Data" version="1.0.2" targetFramework="net472" />
-  <package id="System.Net.Http" version="4.3.4" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
-  <package id="System.Runtime" version="4.3.0" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net472" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.AccessControl" version="6.0.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.ProtectedData" version="4.7.0" targetFramework="net472" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net472" />
-  <package id="System.Security.Permissions" version="6.0.0" targetFramework="net472" />
-  <package id="System.Security.Principal.Windows" version="5.0.0" targetFramework="net472" />
-  <package id="System.Text.Encoding" version="4.3.0" targetFramework="net472" />
-  <package id="System.Text.Encodings.Web" version="6.0.0" targetFramework="net472" />
-  <package id="System.Text.Json" version="4.7.2" targetFramework="net472" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net472" />
-</packages>
